requirements.txt
apache-airflow-providers-postgres
pandas
apache-airflow-providers-google
pyarrow